@@ -1,7 +1,7 @@
 import test from 'ava';
 import execa from 'execa';
 
-test(async t => {
+test('main', async t => {
 	const {stdout} = await execa('./cli.js');
 	t.true(stdout.length > 0);
 });
